requirements.txt
numpy==1.9.0
pytorch==1.7.0
torchvision==0.8.1
pyyaml==5.1